# Add `Trues` type and methods