--- a/src/trues.jl
+++ b/src/trues.jl
@@ -21,17 +21,17 @@ const Trues = Ones{Bool, N, Axes} where {N, Axes}
 const Falses = Zeros{Bool, N, Axes} where {N, Axes}
 
 
-# y[mask] = x when mask isa Trues
+# y[mask] = x when mask isa Trues (cf y[:] = x)
 function Base.setindex!(y::AbstractArray{T,N}, x, mask::Trues{N}) where {T,N}
-	@boundscheck size(x) == size(mask) == size(y) || throw(DimensionMismatch())
-	@boundscheck checkbounds(x, mask)
-	@boundscheck axes(mask) == axes(x) || throw("axes mismatch")
+    @boundscheck size(x) == size(mask) == size(y) || throw(DimensionMismatch())
+    @boundscheck checkbounds(x, mask)
+#   @boundscheck axes(mask) == axes(x) || throw("axes mismatch")
     copyto!(y, x)
 end
 
-# x[mask] when mask isa Trues
+# x[mask] when mask isa Trues (cf x[trues(size(x))] or x[:])
 function Base.getindex(x::AbstractArray{T,D}, mask::Trues{D}) where {T,D}
-	@boundscheck size(x) == size(mask) || throw(DimensionMismatch())
-	@boundscheck checkbounds(x, mask)
-	return vec(x)
+    @boundscheck size(x) == size(mask) || throw(DimensionMismatch())
+    @boundscheck checkbounds(x, mask)
+    return vec(x)
 end