--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1013,3 +1013,15 @@ end
         @test axes(U) == (Base.OneTo(3),Base.OneTo(3))
     end
 end
+
+@testset "Trues" begin
+    @test Trues(2,3) == Trues((2,3)) == trues(2,3)
+    @test Falses(2,3) == Falses((2,3)) == falses(2,3)
+    dim = (4,5)
+    mask = Trues(dim)
+    x = randn(dim)
+    @test x[mask] == vec(x) # getindex
+    y = similar(x)
+    y[mask] = x # setindex!
+    @test y == x
+end