--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1024,4 +1024,7 @@ end
     y = similar(x)
     y[mask] = x # setindex!
     @test y == x
+    @test_throws BoundsError ones(3)[Trues(2)]
+    @test_throws BoundsError setindex!(ones(3), zeros(3), Trues(2))
+    @test_throws DimensionMismatch setindex!(ones(2), zeros(3), Trues(2))
 end