The "show business" (pun intended) at #109 was just a warm-up.  This PR is the one that matters.
Here's a bit of background.  Unaware of `FillArrays.jl`, I created a somewhat similar package [ConstantArrays.jl](https://github.com/JeffFessler/ConstantArrays.jl), and only later learned about the duplication.
One of my main motivations for writing `ConstantArrays` was to have very efficient versions of `getindex` and `setindex!` for "masks" that are `trues(dim)`.
I have timing tests that show the methods for `getindex` and `setindex!` in this PR reduce the time from 1.6ms to 28ns for `getindex` and from 2.3ms to 0.6ms for `setindex!` using `Trues`:
https://github.com/JeffFessler/ConstantArrays.jl/blob/master/test/time.jl
If this PR goes through then I can delete the otherwise redundant `ConstantArrays` and use `FillArrays` hereafter.
BTW, I don't have a use case for `Falses` right now, but I included it for symmetry with `Trues`.
There is one possible test for matching `axes` that I included but left commented out because I don't think it is relevant when indexing using `Bool` arrays, but I am not that familiar with `axes` so I left it there in case one of you experts says it is needed.