@dlfivefifty 2020-08-06
The error messages are still wrong: they need to match `Array` and call `checkbounds` to get the message correct. 

But having thought about this a bit more, I'm not convinced that  overloading `getindex` is the right way to do this. I think we may want to overload `to_indices`, see 

https://github.com/JuliaLang/julia/blob/4c1e3c039775e56c72a4a638db0a8a20f4972dd3/base/multidimensional.jl#L719

That is, perhaps something like this, which just returns the corresponding axis:
```julia
Base.to_indices(A::AbstractArray{N}, inds, I::Tuple{Trues{N}}) where N = Base.to_indices(A, inds, (:,))
```
The benefit is you don't need to worry about bounds checking. The question though is whether there are negative consequences of this.