@JeffFessler 2020-08-06
I need help understanding / fixing the problem with the error messages.
Here are the errors messages for `Trues` and the corresponding messages for `trues` -- they look essentially identical to me:
```
ones(3)[trues(2)]
ERROR: BoundsError: attempt to access 3-element Array{Float64,1} at index [Bool[1, 1]]

ones(3)[Trues(2)]
ERROR: BoundsError: attempt to access 3-element Array{Float64,1} at index [true, true]


setindex!(ones(3), zeros(3), trues(2))
ERROR: BoundsError: attempt to access 3-element Array{Float64,1} at index [Bool[1, 1]]

setindex!(ones(3), zeros(3), Trues(2))
ERROR: BoundsError: attempt to access 3-element Array{Float64,1} at index [true, true]


setindex!(ones(2), zeros(3), trues(2))
ERROR: DimensionMismatch("tried to assign 3 elements to 2 destinations")

setindex!(ones(2), zeros(3), Trues(2))
ERROR: DimensionMismatch("tried to assign 3 elements to 2 destinations")
```

Sorry, but I do not see how to use `to_indices` here because the example you show uses a `Tuple` of 1D `Trues` but the method I need uses a single `Trues` argument, like `ones(3)[Trues(3)]`, not a Tuple.  Nevertheless, I tried the suggestion, but my timing tests showed that it did not provide any speedup over how `getindex` is handled by the existing code in `FillArrays`.  If interested you can see the details in this test/timing file:
https://github.com/JeffFessler/ConstantArrays.jl/blob/master/test/to_indices.jl

Perhaps your concern has to do with types that use non-`OneTo` axes.  I need this extension only for arrays with `OneTo` axes.  The only non-`OneTo` axes I recall ever using are those in `FFTViews` and I did a little test with one of those types and found that it did not support using an array of Bools as indices.  (The little test is in the comments at the bottom of the link above.)

Thus, I've pushed a new commit where I've modified the `getindex` method to work *only* with `OneTo` types, and revert to whatever existing `getindex` method there is (if any) for other types.  This simplifies the bounds checking to just a size check.  If this restriction to `OneTo` types looks acceptable to you then I will modify `setindex!` with the same strategy.

I don't understand enough about axes to try to generalize this to non-`OneTo` types anyway, especially when it comes to using Bool indices for such types.