@dlfivefifty 2020-08-06
OK we can forget `to_indices` for now, just thought it might be cleaner. I've added a comment of the error message that looks wrong