@JeffFessler 2020-08-06
Ok, I have updated `setindex!` so that it is also applicable solely to `OneTo` cases, which seems cleaner, and I've timed it and it still provides the performance benefit expected.