@dlfivefifty 2020-08-07
You still need to fix the errors for `getindex(A::$Typ{T}, kr::AbstractVector{Bool}) where T`