@JeffFessler 2020-08-07
I had replied to that request right below where you suggested it, and then I clicked "resolve conversation" - which in retrospect probably was another freshman mistake on my part.  It's unclear to me who is supposed to click it.  Now I've "unresolved it" so perhaps my previous reply will be more visible.

The short version of my reply is "not my job because that was not code that I added and I shouldn't have messed with it in the first place in this PR."  So I reverted it back to the original version that had no string. 

Now the weird thing is that github is not displaying correctly the fact that I reverted it in the "Files changed" tabl.
If you go directly to the file you will see that the non-compliant string is gone:
https://github.com/JeffFessler/FillArrays.jl/blob/63968e46b2098832497924874e1aef62dee16d19/src/FillArrays.jl#L162
But (at least for me) under "Files changed" the difference does not reflect the fact that I reverted it.