@JeffFessler 2020-08-07
Thanks for the merge.  I am happy to do a "service" when I am confident about the code, like I did with the docstrings, but I am not confident about @inbounds and @boundscheck details.